# docs/openapi.yaml
openapi: 3.0.3
info:
  title: cardinal-wings API
  description: >
    REST API daemon for managing cardinal over HTTP. Panel-oriented surface
    over cardinal's Docker-compatible API, with multi-node routing (`?node=`),
    async tasks, streaming logs/exec, metrics and role-based auth.
  version: "0.2.0-dev"
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
...
paths:
...
  /v1/containers/{id}/sftp:
    get:
      tags: [containers]
      summary: SFTP credential status for a container
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: status } }
    put:
      tags: [containers]
      summary: Set SFTP credential for a container (admin)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
      responses: { "200": { description: set } }
    delete:
      tags: [containers]
      summary: Remove SFTP credential (admin)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: removed } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...